fix(layout): address PR review on header and icon changes
- Replace House icon with DotsNine in FeaturesMenu
- Replace Info icon with Question in HowItWorksModal
- Remove Header from shared route layout (route.tsx)
- Restore Header to onboarding.tsx and add it to each feature page
- Each page now controls its own Header, Stepper step, and right section
