+++ src/renderer/src/components/features-menu.tsx
@@ -2,7 +2,7 @@ import { FEATURES } from "@/constants";
 import { Grid, Stack, styled } from "@/styled/jsx";
 import type { FeatureFlowEnum } from "@/types/features";
 import { Link } from "@tanstack/react-router";
-import { House } from "phosphor-react";
+import { DotsNine } from "phosphor-react";
 import FeatureIcon from "./feature-icon";
 import Button from "./ui/button";
 import Card from "./ui/card";
@@ -14,7 +14,7 @@ export default function FeaturesMenu() {
     <Popover>
       <PopoverTrigger asChild>
         <Button size="icon-sm" style={{ padding: 2 }} aria-label="Ir al inicio">
-          <House size={32} />
+          <DotsNine size={32} />
         </Button>
       </PopoverTrigger>
       <PopoverContent align="end">
