+++ src/renderer/src/routes/app.$feature/onboarding.tsx
@@ -5,9 +5,13 @@ import {
 } from "@tanstack/react-router";
 
 import DropArea from "@/components/drop-area";
+import FeaturesMenu from "@/components/features-menu";
 import HiddenInput from "@/components/hidden-input";
 import HowItWorks from "@/components/how-it-works";
+import HowItWorksModal from "@/components/how-it-works-modal";
+import Stepper from "@/components/home/stepper";
 import Footer from "@/components/layout/footer";
+import Header from "@/components/layout/header";
 import MainContent from "@/components/layout/main-content";
 import BackButton from "@/components/ui/back-button";
 import Button from "@/components/ui/button";
@@ -16,6 +20,7 @@ import { SectionTitle } from "@/layout/section-title";
 import { addFiles } from "@/reducers/file/actions";
 import { useSetTutorialSeen, useTutorialSeen } from "@/store/useLocal";
 import { HStack, Stack, styled } from "@/styled/jsx";
+import { FeatureFlowEnum, featureName } from "@/types/features";
 import { useRef } from "react";
 
 // FIRST step of the processing workflow
@@ -55,6 +60,20 @@ function RouteComponent() {
 
   return (
     <>
+      <Header
+        title={featureName(feature)}
+        center={
+          feature === FeatureFlowEnum.Dataset ? (
+            <Stepper currentStep={1} />
+          ) : undefined
+        }
+        right={
+          <HStack>
+            {tutorialSeen && <HowItWorksModal feature={feature} />}
+            <FeaturesMenu />
+          </HStack>
+        }
+      />
       <MainContent>
         {tutorialSeen ? (
           <Stack gap="8">
