+++ src/renderer/src/routes/app.$feature/route.tsx
@@ -2,18 +2,12 @@ import {
   Outlet,
   createFileRoute,
   redirect,
-  useLocation,
 } from "@tanstack/react-router";
 import { z } from "zod";
 
-import FeaturesMenu from "@/components/features-menu";
-import Stepper from "@/components/home/stepper";
-import HowItWorksModal from "@/components/how-it-works-modal";
-import Header from "@/components/layout/header";
 import FileProvider from "@/context/File";
-import { useTutorialSeen } from "@/store/useLocal";
-import { HStack, Stack } from "@/styled/jsx";
-import { FeatureFlowEnum, featureName } from "@/types/features";
+import { Stack } from "@/styled/jsx";
+import { FeatureFlowEnum } from "@/types/features";
 
 // Validation schema for feature parameter
 const featureParamSchema = z.object({
@@ -36,32 +30,8 @@ export const Route = createFileRoute("/app/$feature")({
 });
 
 function AppLayoutRoute() {
-  const { feature } = Route.useParams();
-  const location = useLocation();
-  const tutorialSeen = useTutorialSeen(feature);
-
-  let currentStep = 1;
-  const path = location.pathname;
-  if (path.includes("/process")) currentStep = 2;
-  else if (path.includes("/validation")) currentStep = 3;
-  else if (path.includes("/finish")) currentStep = 4;
-
   return (
     <Stack width="screen" height="screen" gap="0">
-      <Header
-        title={featureName(feature)}
-        center={
-          feature === FeatureFlowEnum.Dataset ? (
-            <Stepper currentStep={currentStep} />
-          ) : undefined
-        }
-        right={
-          <HStack>
-            {tutorialSeen && <HowItWorksModal feature={feature} />}
-            <FeaturesMenu />
-          </HStack>
-        }
-      />
       <FileProvider>
         <Outlet />
       </FileProvider>
